Revert "sphinx label warnings"
This reverts commit d9e6b4f51bc8517453653be2904ab6db9aaab85e.

It was able to remove the warnings about duplicate labels, but then it
shows [WA1] instead of [1] in the generated doc. And for things cited on
multiple pages, it uses the same everywhere, so on a single page, you
can have a mix of [I1], [WI2], etc.
Not very pretty.