Fewer copies and no GIL for hera
Now the input arrays are not copied as long as they use a float64 data
type, even if they are not contiguous. That's not important here, but I
wanted an example of how to do it.
More importantly, no need to hold the GIL.
I was too lazy to benchmark to see if that changed anything...