Build python first
This subdirectory is built sequentially and takes a very long time.
Starting it first, in a parallel build, it is still the last to end, but
since it started earlier it also ends earlier.

Usually I build the plugin directly with
  python3 setup.py build_ext --inplace -j20
But I don't think there is a natural way to do that from the makefile
(except for the unlimited -j), since I doubt setuptools knows how to
cooperate with make, ninja, etc to share a pool of workers.