Add support for irregular (non-evenly, completely sampled) Betti curves.
There's a conceptual difference in the meaning of the constructor's
argument when compared to the existing BettiCurve class. In
IrregularBettiCurve, only intervals born in [x_min, x_max) contribute
to the Betti curve. The behavior of BettiCurve is of less use when not
sampling evenly.

TODO: We might want to move this "cropping" of the persistence diagram
to a more general class that transforms PDs, instead of arbitrarily
having it as part of the IrregularBettiCurve class.