Documentation: Doxygen warnings
When running doxygen on the sources we get (a.o.) the following warnings:
```
.../gudhi-devel/src/Contraction/include/gudhi/Edge_contraction.h:214: warning: unbalanced grouping commands
.../gudhi-devel/src/common/doc/examples.h:3: warning: the name 'Examples' supplied as the argument in the \file statement is not an input file
error: bib file .../gudhi-devel/biblio/how_to_cite_gudhi.bib not found!
.../gudhi-devel/src/common/doc/installation.h:271: warning: included file biblio/how_to_cite_gudhi.bib is not found. Check your EXAMPLE_PATH
```
This is corrected in the different files and in the settings in src/Doxyfile.in

Also there are a number of settings that are obsolete like:
```
warning: Tag 'LATEX_SOURCE_CODE' at line 1835 of file '.../gudhi-devel/build_windows_fork/Doxyfile' has become obsolete.
```
This is corrected in the src/Doxyfile.in file