Documentation: Non persistent links
During the link check we found some references to `#title..` in respect to links to CGAL, these links work at the moment but are, by CGAL, generated links during runtime based on the HTML `<H...>` tags.
When a new paragraph or chapter is added these numbers will change and thus will the link point to the wrong place. All these items have also a more persistent reference anchor that is generated based on the `<name>` of the section (e.g. `\section <name> <description>`).
Therefore it is better to used the references based on the `<name>`.