Documentation: handling references to examples in a more transparent way
In case we have the files`Bitmap_cubical_complex/Random_bitmap_cubical_complex.cpp` this will be linked to `_random_bitmap_cubical_complex_8cpp-example.html` and this is now hard coded.
When the setting of `CASE_SENSE_NAMES` is set to `YES` the mapping is to `Random_bitmap_cubical_complex_8cpp-example.html` but the hard coded link is still to `_random_bitmap_cubical_complex_8cpp-example.html` so the link cannot be resolved (at runtime!). Same could happen when the `CASE_SENSE_NAMES` is left blank and the the value is system dependent. It is unlikely that this happens but ...

By introducing the `ALIASES` `gudhi_example_link` this can be handled by doxygen in a more transparent way.

Furthermore wit 2 examples there was a double `cpp` in the text part, this has been corrected as well

Notre this solution supersedes #621 and maked #621 obsolete.